+++ pwiz_tools/Shared/Common/CommonUtil.csproj
@@ -141,6 +141,7 @@
     <Compile Include="SystemUtil\CenterWinDialog.cs" />
     <Compile Include="SystemUtil\CommonException.cs" />
     <Compile Include="SystemUtil\MemoryBlockingCollection.cs" />
+    <Compile Include="SystemUtil\ParallelEx.cs" />
     <Compile Include="SystemUtil\TrackAttribute.cs" />
     <Compile Include="SystemUtil\CommonActionUtil.cs" />
     <Compile Include="SystemUtil\ValueCache.cs" />
