+++ pwiz_tools/Skyline/Util/Util.cs
@@ -2128,118 +2128,6 @@ namespace pwiz.Skyline.Util
         }
     }
 
-    public static class ParallelEx
-    {
-        // This can be set to true to make debugging easier.
-        public static readonly bool SINGLE_THREADED = false;
-
-        private class IntHolder
-        {
-            public IntHolder(int theInt)
-            {
-                TheInt = theInt;
-            }
-
-            public int TheInt { get; private set; }
-        }
-
-        public static int GetThreadCount(int? maxThreads = null)
-        {
-            if (SINGLE_THREADED)
-                return 1;
-            int threadCount = Environment.ProcessorCount;
-            int maxThreadCount = maxThreads ?? 8; // Trial with maximum of 8
-            if (threadCount > maxThreadCount)
-                threadCount = maxThreadCount;
-            return threadCount;
-        }
-
-        public static void For(int fromInclusive, int toExclusive, Action<int> body, Action<AggregateException> catchClause = null, int? maxThreads = null)
-        {
-            Action<int> localBody = i =>
-            {
-                LocalizationHelper.InitThread(); // Ensure appropriate culture
-                body(i);
-            };
-            LoopWithExceptionHandling(() =>
-            {
-                using (var worker = new QueueWorker<IntHolder>(null, (h, i) => localBody(h.TheInt)))
-                {
-                    worker.RunAsync(GetThreadCount(maxThreads), typeof(ParallelEx).Name);
-                    for (int i = fromInclusive; i < toExclusive; i++)
-                    {
-                        if (worker.Exception != null)
-                            break;
-                        worker.Add(new IntHolder(i));
-                    }
-                    worker.DoneAdding(true);
-                    if (worker.Exception != null)
-                        throw new AggregateException(@"Exception in Parallel.For", worker.Exception);   
-                }
-            }, catchClause);
-//            LoopWithExceptionHandling(() => Parallel.For(fromInclusive, toExclusive, PARALLEL_OPTIONS, localBody), catchClause);
-        }
-
-        public static void ForEach<TSource>(IEnumerable<TSource> source, Action<TSource> body, Action<AggregateException> catchClause = null, int? maxThreads = null) where TSource : class
-        {
-            Action<TSource> localBody = o =>
-            {
-                LocalizationHelper.InitThread(); // Ensure appropriate culture
-                body(o);
-            };
-            LoopWithExceptionHandling(() =>
-            {
-                using (var worker = new QueueWorker<TSource>(null, (s, i) => localBody(s)))
-                {
-                    worker.RunAsync(GetThreadCount(maxThreads), typeof(ParallelEx).Name);
-                    foreach (TSource s in source)
-                    {
-                        if (worker.Exception != null)
-                            break;
-                        worker.Add(s);
-                    }
-                    worker.DoneAdding(true);
-                    if (worker.Exception != null)
-                        throw new AggregateException(@"Exception in Parallel.ForEx", worker.Exception); 
-                }
-            }, catchClause);
-//            LoopWithExceptionHandling(() => Parallel.ForEach(source, PARALLEL_OPTIONS, localBody), catchClause);
-        }
-
-        private static void LoopWithExceptionHandling(Action loop, Action<AggregateException> catchClause)
-        {
-            try
-            {
-                loop();
-            }
-            catch (AggregateException x)
-            {
-                Exception ex = null;
-                x.Handle(inner =>
-                {
-                    if (inner is OperationCanceledException)
-                    {
-                        if (!(ex is OperationCanceledException))
-                            ex = inner;
-                        return true;
-                    }
-                    if (catchClause == null)
-                    {
-                        if (ex == null)
-                            ex = inner;
-                        return true;
-                    }
-                    return false;
-                });
-
-                if (ex != null)
-                    Helpers.WrapAndThrowException(ex);
-                if (catchClause != null)
-                    catchClause(x);
-            }
-        }
-    }
-
     public class Alarms
     {
         private readonly Dictionary<object, AlarmInfo> _timers =
